ERILON-4 Поправить отображение вероятностей выпадения врожденки на странице Врожденных особенностей
ERILON-7 Добавить во врожденки возможность задавать/роллить числовое значение

+++ models/raceInherent.js
@@ -5,7 +5,7 @@
 
 module.exports = function(sequelize, DataTypes) {
     var RaceInherent = sequelize.define("RaceInherent", {
-        race_probability: DataTypes.DOUBLE
+        race_probability: DataTypes.INTEGER
     }, {
         classMethods: {
             associate: function(models) {
